+++ tabbycat/importer/importers/anorak.py
@@ -296,9 +296,22 @@ class AnorakTournamentDataImporter(BaseTournamentDataImporter):
 
     def import_motions(self, f):
         motions_interpreter = make_interpreter(
-            round=lambda x: tm.Round.objects.lookup(x, tournament=self.tournament),
+            DELETE=['rounds', 'seq']
         )
-        self._import(f, mm.Motion, motions_interpreter)
+        motions = self._import(f, mm.Motion, motions_interpreter)
+
+        def round_motions_interpreter(lineno, line):
+            if not line.get('rounds'):
+                return
+            motion = motions[lineno]
+            for round_name, seq in zip(line['rounds'].split(";"), line['seq'].split(";")):
+                seq = seq or 1
+                yield {
+                    'seq'    : seq,
+                    'motion' : motion,
+                    'round'  : tm.Round.objects.get(abbreviation=round_name, tournament=self.tournament),
+                }
+        self._import(f, mm.RoundMotions, round_motions_interpreter)
 
     def import_sides(self, f):
         def side_interpreter(lineno, line):