+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Change Log
 ------------------
 *Release date: TBD*
 
+- Made it possible to re-use an existing motion for multiple rounds.
 - Added the "average individual speaker score" metric which averages the scores of all substantive speeches by the team within preliminary rounds.
 - Renamed the "average speaker score" to "average total speaker score."
 - Implemented a new server architecture on Heroku that should significantly improve performance under load. If upgrading an existing Heroku instance this requires a few tweaks: